Add generic discovery of the submodules GIT server
The GIT modules are set up with the Gitlab Token based on the same URI
of the repository server in use. This allows Enterprise Gitlab, with
domains other than `gitlab.com` to be initialized.

GIT modules with URL to other servers are not changed, but should be
initialize well if it doesn't require authentication.

GIT modules located in other servers which requires authentication won't
be initialized and most likely will fail.